Moved weights out of examples folder
